Getting more current code in place.

+++ ground/openpilotgcs/src/libs/utils/coordinateconversions.h
@@ -41,7 +41,7 @@ public:
     CoordinateConversions();
     int NED2LLA_HomeECEF(double BaseECEFcm[3], double NED[3], double position[3]);
     int NED2LLA_HomeLLA(double LLA[3], double NED[3], double position[3]);
-    void RneFromLLA(double LLA[3], double Rne[3][3]);
+    void RneFromLLA(double LLA[3], float Rne[3][3]);
     void LLA2ECEF(double LLA[3], double ECEF[3]);
     int ECEF2LLA(double ECEF[3], double LLA[3]);
     void LLA2Base(double LLA[3], double BaseECEF[3], float Rne[3][3], float NED[3]);
